Added task 1 fix

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@ void print_binary(unsigned long int n)
 {
 	int first_bit_1_printed = 0;
 	int num_shifts;
-	int bit;
+	unsigned long int bit;
 
 	if (n == 0)
 	{
